@@ -12,28 +12,27 @@
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@iconify/json": "^2.2.162",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/eslint": "8.56.0",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
-		"eslint": "^8.56.0",
+		"@iconify/json": "^2.2.290",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.15.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@types/eslint": "^9.6.1",
+		"@typescript-eslint/eslint-plugin": "^8.18.2",
+		"@typescript-eslint/parser": "^8.18.2",
+		"eslint": "^9.17.0",
 		"eslint-config-prettier": "^9.1.0",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"sass": "^1.69.5",
-		"svelte": "^5.0.0-next.33",
-		"svelte-check": "^3.6.0",
-		"svelte-preprocess": "^5.1.3",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.2",
+		"sass": "^1.83.0",
+		"svelte": "^5.16.0",
+		"svelte-check": "^3.8.6",
+		"tslib": "^2.8.1",
+		"typescript": "^5.7.2",
+		"vite": "^6.0.6"
 	},
 	"type": "module",
 	"dependencies": {
 		"ical.js": "^1.5.0",
-		"unplugin-icons": "^0.18.1"
+		"unplugin-icons": "^0.18.5"
 	}
 }
